server: move StatementDetails funcs to own file
Previously, the StatementDetails api were located in the file
`combined_statement_stats.go`. This file is quite large. This
commit extracts StatementDetails api functions into its own file
`statement_details.go` for improved organization.

Epic: none

Release note: None
